pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "finwin"
version = "0.1.0"
description = "Financial context collector for stocks"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "Finwin Team" }
]
keywords = ["finance", "stocks", "news", "context", "yfinance"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "pydantic>=2.0",
    "httpx>=0.25",
    "feedparser>=6.0",
    "beautifulsoup4>=4.12",
    "trafilatura>=1.6",
    "yfinance>=0.2",
    "pdfminer.six>=20221105",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
    "ruff>=0.1",
]

[project.scripts]
finwin = "finwin.cli:main"

[project.urls]
Homepage = "https://github.com/skylarkgit/finwin"
Repository = "https://github.com/skylarkgit/finwin"

[tool.hatch.build.targets.wheel]
packages = ["finwin"]

[tool.ruff]
line-length = 88
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
ignore = ["E501"]  # Line too long - handled by formatter

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
